[ Style ] domian/view 폴더 생성 및 하위문서 이동
