--- src/Readline.js
@@ -1,8 +1,0 @@
-const readline = require("readline")
-
-const RL = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-})
-
-module.exports = RL;
