+++ src/domain/Constant_domain.js
@@ -1,0 +1,12 @@
+const RACING = {
+    RANDOM_MIN: 0,
+    RANDOM_MAX: 9,
+    GO_POINT: 4,
+
+    CAR_NAME_MIN: 0,
+    CAR_NAME_MAX: 5,
+
+    ROUND_MIN: 1,
+};
+
+module.exports = { RACING };
