+++ src/domain/Random.js
@@ -1,8 +1,7 @@
-
 const Random = {
     getRandomNumber(min, max) {
         return Math.round(Math.random() * (max - min) + min);
-    }
-}
+    },
+};
 
 module.exports = Random;
