+++ src/view/Readline.js
@@ -1,0 +1,8 @@
+const readline = require("readline");
+
+const RL = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+});
+
+module.exports = RL;
